Fixed issue with nim stripping out `"` characters automatically.
If there is a quoted parameter, nim will strip out the quote characters, so user commands end up breaking.  It's safe to assume any parameter that has a space was quoted, since it would end up as 2 parameters if it wasn't, so this fix reintroduces the quotes if a space is found within ANY parameter in the sequence.